add decimal for weight/height/girth and adds initial values for history

+++ src/MyFoodDoc.App.Application/Payloads/User/UpdateUserPayload.cs
@@ -11,7 +11,7 @@ namespace MyFoodDoc.App.Application.Payloads.User
 
         public Gender Gender { get; set; }
 
-        public int Height { get; set; }
+        public decimal Height { get; set; }
 
         public int InsuranceId { get; set; }
